backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2
requests
google-generativeai
python-dotenv
